Finished mobile and desktop builder input for tile/entity placement.
