+++ Assets/Resources/Prefabs/Builder/Ripple.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 1892960647589689741}
   - component: {fileID: 2000613154666451814}
   - component: {fileID: 3712415924894535025}
+  - component: {fileID: 6947966270295779050}
   m_Layer: 0
   m_Name: Ripple
   m_TagString: Untagged
@@ -75,3 +76,15 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!225 &6947966270295779050
+CanvasGroup:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7620636519634441475}
+  m_Enabled: 1
+  m_Alpha: 1
+  m_Interactable: 1
+  m_BlocksRaycasts: 1
+  m_IgnoreParentGroups: 0
