--- Assets/Scripts/Input/BuilderInputHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cdcde9cd96eeb724c932eaeea1f50c1c
